renaming to use-strict-null-equality-checks
